+++ TTPLSite.New/Pages/Shared/_Layout.cshtml
@@ -146,59 +146,63 @@
                                             <div class="menu-sub-box">
                                                 <h3>Front End</h3>
                                                 <li><a class="dropdown-item"
-                                                        href="/angular-technologies">Angular</a></li>
+                                                   href="/technologies/angular-technologies">Angular</a>
+                                            </li>
                                                 <li><a class="dropdown-item"
-                                                        href="/reactjs-technologies">Reactjs</a></li>
-                                                <li><a class="dropdown-item" href="/vuejs-technologies">Vue.js</a>
+                                                   href="/technologies/reactjs-technologies">Reactjs</a>
+                                            </li>
+                                            <li>
+                                                <a class="dropdown-item" href="/technologies/vuejs-technologies">Vue.js</a>
                                                 </li>
-                                                <li><a class="dropdown-item" href="/uiux-technologies">UI/UX</a>
+                                            <li>
+                                                <a class="dropdown-item" href="/technologies/uiux-technologies">UI/UX</a>
                                                 </li>
                                             </div>
                                             <div class="menu-sub-box">
                                                 <h3>Back End</h3>
                                 <li>
-                                    <a class="dropdown-item" href="/ruby-on-rails-technologies">
+                                    <a class="dropdown-item" href="/technologies/ruby-on-rails-technologies">
                                         Ruby
                                                         on Rails</a>
                                                 </li>
                                 <li>
-                                    <a class="dropdown-item" href="/nodejs-technologies">Node.js</a>
+                                    <a class="dropdown-item" href="/technologies/nodejs-technologies">Node.js</a>
                                                 </li>
                                                 <li><a class="dropdown-item"
-                                                        href="technologies/tegolang.html">Golang</a></li>
+                                                        href="/technologies/tegolang.html">Golang</a></li>
                                                 <li><a class="dropdown-item"
-                                       href="/laravel-technologies">Laravel</a>
+                                       href="/technologies/laravel-technologies">Laravel</a>
                                 </li>
-                                <li><a class="dropdown-item" href="/net-technologies">.NET</a></li>
-                                <li><a class="dropdown-item" href="/java-technologies">Java</a></li>
+                                <li><a class="dropdown-item" href="/technologies/net-technologies">.NET</a></li>
+                                <li><a class="dropdown-item" href="/technologies/java-technologies">Java</a></li>
                                 <li>
-                                    <a class="dropdown-item" href="/python-technologies">Python</a>
+                                    <a class="dropdown-item" href="/technologies/python-technologies">Python</a>
                                                 </li>
                                             </div>
                                             <div class="menu-sub-box">
                                                 <h3>QA</h3>
                                                 <li><a class="dropdown-item"
-                                       href="/automation-testing-technologies">
+                                       href="/technologies/automation-testing-technologies">
                                         Automation
                                                         Testing</a></li>
                                                 <li><a class="dropdown-item"
-                                       href="/software-testing-technologies">
+                                       href="/technologies/software-testing-technologies">
                                         Software
                                                         Testing</a></li>
                                             </div>
                                             <div class="menu-sub-box">
                                                 <h3>Cloud</h3>
                                 <li>
-                                    <a class="dropdown-item" href="/azure-technologies">Azure</a>
+                                    <a class="dropdown-item" href="/technologies/azure-technologies">Azure</a>
                                                 </li>
-                                <li><a class="dropdown-item" href="/aws-technologies">AWS</a></li>
+                                <li><a class="dropdown-item" href="/technologies/aws-technologies">AWS</a></li>
                                 <li>
-                                    <a class="dropdown-item" href="/devops-technologies">DevOps</a>
+                                    <a class="dropdown-item" href="/technologies/devops-technologies">DevOps</a>
                                                 </li>
                                             </div>
                                         </ul>
                                     </li>
-                                    <li class="nav-item"> <a class="nav-link" href="/portfolio"> Portfolio</a> </li>
+                            <li class="nav-item"> <a class="nav-link" href="/technologies/portfolio"> Portfolio</a> </li>
                                     <!--<li class="nav-item"> <a class="nav-link" href="career.html"> Career</a> </li>-->
                                     <!--li class="nav-item dropdown"> <a class="nav-link dropdown-toggle" href="#" role="button" data-bs-toggle="dropdown"
                                                   aria-expanded="false"> Case Study </a>
@@ -213,7 +217,7 @@
                                         <li><a class="dropdown-item" href="#">One Detail page</a></li>
                                       </ul>
                                     </li-->
-                                    <li class="nav-item"> <a class="nav-link" href="/contact-us"> Contact us</a> </li>
+                            <li class="nav-item"> <a class="nav-link" href="/technologies/contact-us"> Contact us</a> </li>
                                 </ul>
                             </div>
                         </div>
