+++ TTPLSite.New/Pages/custom-development-services.cshtml
@@ -2,7 +2,6 @@
 @model TTPLSite.New.Pages.custom_development_servicesModel
 @{
 }
-
 <div class="services-saction-banner app-mobile">
     <div class="container">
         <div class="row">
