+++ TTPLSite.New/Pages/mobile-app-development-services.cshtml
@@ -2,7 +2,6 @@
 @model TTPLSite.New.Pages.mobile_app_developmentModel
 @{
 }
-
 <div class="services-saction-banner mobile-app">
     <div class="container">
         <div class="row">
