+++ TTPLSite.New/Pages/our-leadership.cshtml
@@ -2,7 +2,6 @@
 @model TTPLSite.New.Pages.our_leadershipModel
 @{
 }
-
 <div class="services-saction-banner leadership">
     <div class="container">
         <div class="row">
@@ -251,3 +250,22 @@
         </div>
     </div>
 </div>
+
+<!-- /// CONSULTATION-Section-start /// -->
+<!--<div class="CONSULTATION-section">
+  <div class="container">
+    <div class="row">
+      <div class="col-md-5 alting">
+        <div class="NEED-CONSU"> <img src="images/clt.png" alt="" /> </div>
+      </div>
+      <div class="col-md-7">
+        <div class="CONSULTATION-cont">
+          <h2 class="text-org">We are here to answer your questions 24/7</h2>
+          <h3 class="our-text altd">NEED A CONSULTATION?</h3>
+          <div class="btn-2 sub"> <a href="">Contact Us</a> </div>
+        </div>
+      </div>
+    </div>
+  </div>
+</div>-->
+<!-- /// CONSULTATION-Section-end /// -->
