+++ TTPLSite.New/Pages/our-philosophy.cshtml
@@ -2,7 +2,6 @@
 @model TTPLSite.New.Pages.our_philosophyModel
 @{
 }
-
 <div class="services-saction-banner our-philosophy">
     <div class="container">
         <div class="row">
