+++ TTPLSite.New/Pages/portfolio.cshtml
@@ -2,7 +2,6 @@
 @model TTPLSite.New.Pages.portfolioModel
 @{
 }
-
 <div class="services-saction-banner portfolio">
     <div class="container">
         <div class="row">
