+++ TTPLSite.New/Pages/technologies/azure-technologies.cshtml.cs
@@ -1,9 +1,9 @@
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.RazorPages;
 
-namespace TTPLSite.New.Pages
+namespace TTPLSite.New.Pages.technologies
 {
-    public class software_testing_technologiesModel : PageModel
+    public class azure_technologiesModel : PageModel
     {
         public void OnGet()
         {
