+++ TTPLSite.New/Pages/technologies/devops-technologies.cshtml.cs
@@ -1,9 +1,9 @@
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.RazorPages;
 
-namespace TTPLSite.New.Pages
+namespace TTPLSite.New.Pages.technologies
 {
-    public class reactjs_technologiesModel : PageModel
+    public class devops_technologiesModel : PageModel
     {
         public void OnGet()
         {
