+++ TTPLSite.New/Pages/technologies/laravel-technologies.cshtml.cs
@@ -1,9 +1,9 @@
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.RazorPages;
 
-namespace TTPLSite.New.Pages
+namespace TTPLSite.New.Pages.technologies
 {
-    public class ruby_on_rails_technologiesModel : PageModel
+    public class laravel_technologiesModel : PageModel
     {
         public void OnGet()
         {
