+++ TTPLSite.New/Pages/technologies/net-technologies.cshtml.cs
@@ -1,9 +1,9 @@
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.RazorPages;
 
-namespace TTPLSite.New.Pages
+namespace TTPLSite.New.Pages.technologies
 {
-    public class java_technologiesModel : PageModel
+    public class net_technologiesModel : PageModel
     {
         public void OnGet()
         {
