+++ TTPLSite.New/Pages/technologies/nodejs-technologies.cshtml.cs
@@ -1,9 +1,9 @@
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.RazorPages;
 
-namespace TTPLSite.New.Pages
+namespace TTPLSite.New.Pages.technologies
 {
-    public class laravel_technologiesModel : PageModel
+    public class nodejs_technologiesModel : PageModel
     {
         public void OnGet()
         {
