+++ TTPLSite.New/Pages/technologies/reactjs-technologies.cshtml.cs
@@ -1,9 +1,9 @@
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.RazorPages;
 
-namespace TTPLSite.New.Pages
+namespace TTPLSite.New.Pages.technologies
 {
-    public class python_technologiesModel : PageModel
+    public class reactjs_technologiesModel : PageModel
     {
         public void OnGet()
         {
