+++ TTPLSite.New/Pages/technologies/ruby-on-rails-technologies.cshtml.cs
@@ -1,9 +1,9 @@
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.RazorPages;
 
-namespace TTPLSite.New.Pages
+namespace TTPLSite.New.Pages.technologies
 {
-    public class vuejs_technologiesModel : PageModel
+    public class ruby_on_rails_technologiesModel : PageModel
     {
         public void OnGet()
         {
