+++ TTPLSite.New/Pages/technologies/vuejs-technologies.cshtml.cs
@@ -1,9 +1,9 @@
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.RazorPages;
 
-namespace TTPLSite.New.Pages
+namespace TTPLSite.New.Pages.technologies
 {
-    public class net_technologiesModel : PageModel
+    public class vuejs_technologiesModel : PageModel
     {
         public void OnGet()
         {
