+++ TTPLSite.New/Pages/ui-ux-development-services.cshtml
@@ -2,8 +2,6 @@
 @model TTPLSite.New.Pages.ui_ux_development_servicesModel
 @{
 }
-
-<!-- banner-section-start -->
 <div class="services-saction-banner-UX-Design">
     <div class="container">
         <div class="row">
