+++ TTPLSite.New/Pages/why-choose-us.cshtml
@@ -2,8 +2,6 @@
 @model TTPLSite.New.Pages.why_choose_usModel
 @{
 }
-
-
 <div class="services-saction-banner why-chose-us">
     <div class="container">
         <div class="row">
@@ -236,7 +234,7 @@
             <div class="col-sm-12 col-md-6 col-lg-4 pl-0 pr-0">
                 <div class="services-single-items">
                     <div class="services-single-content">
-                        <div class="services-single-icon"> <img src="images/company/info-tech-icon1.png"> </div>
+                        <div class="services-single-icon"> <img src="images/company/web-apps-development.png"> </div>
                         <div class="services-content-inner">
                             <h2>Web Apps Development</h2>
                             <p>
@@ -346,3 +344,21 @@
         </div>
     </div>
 </div>
+
+
+<!--<div class="CONSULTATION-section">
+  <div class="container">
+    <div class="row">
+      <div class="col-md-5 alting">
+        <div class="NEED-CONSU"> <img src="images/clt.png" alt=""> </div>
+      </div>
+      <div class="col-md-7">
+        <div class="CONSULTATION-cont">
+          <h2 class="text-org">We are here to answer your questions 24/7</h2>
+          <h3 class="our-text altd">NEED A CONSULTATION?</h3>
+          <div class="btn-2 sub"> <a href="">Contact Us</a> </div>
+        </div>
+      </div>
+    </div>
+  </div>
+</div>-->
