+++ TTPLSite.New/Pages/work-methodology.cshtml
@@ -2,8 +2,6 @@
 @model TTPLSite.New.Pages.work_methodologyModel
 @{
 }
-
-
 <div class="services-saction-banner methodology">
     <div class="container">
         <div class="row">
@@ -18,7 +16,6 @@
         </div>
     </div>
 </div>
-
 
 <!-- {{{ Work Methodology Start }}} -->
 
@@ -96,7 +93,7 @@
             <li>
                 <div class="direction-r">
                     <div class="flag-wrapper">
-                        <span class="hexa"></span> <img src="images/company/acrol-icon-3.png"> <span class="root-icon">03</span> <span class="flag">Project Kick-off & Team Introduction</span>
+                        <span class="hexa"></span> <img src="images/company/acrol-icon-3-new.png"> <span class="root-icon">03</span> <span class="flag">Project Kick-off & Team Introduction</span>
                     </div>
                     <div class="desc">
                         The scope of our given services and the agreement entail determining and documenting a
@@ -281,4 +278,3 @@
     </div>
   </div>
 </div>-->
-<!-- {{{ Work Methodology End }}} -->
